yukon: custombootimg.mk: use a standard boot image
We've grown used to egregious hacks this century. It's hard to
dazzle us. But for nearly 5 years custombootimg.mk has been doing
just that. We've grown used to the idea of packing recovery into
the boot image, and perhaps we forget that we've been doing it wrong.

We're still pioneers. We're still finding better ways of doing
things. And now with real boot loaders for Open Devices, it's
time to say farewell to an old friend. The future doesn't belong
to unsustainable hacks; it belongs to clean, open code. These new
boot loaders are pulling us into the future, and we'll continue
to follow them.

Change-Id: I6fece1831062967c2d3ae930d472baa8d9b4979e
Signed-off-by: Adam Farden <adam@farden.cz>