Update patch set 2
Patch Set 2: Code-Review-2

Not gonna happen Jake, we don't have any tools to revive the device during a brick, and we don't know how does Sony update bootloaders on device as they don't use the generic FOTA flow.

We will add an assert, send out a PSA and such but won't be adding bootloaders.

Marin, what would be the point in that? Doesn't matter anyways, and looks cleaner this way :)


Blocking this before we what was said above.

Patch-set: 2